backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
langchain-huggingface
sentence-transformers
faiss-cpu
flask
flask-cors
python-dotenv
torch
transformers
accelerate
unstructured
pypdf
pymupdf
gpt4all
llama-cpp-python
vosk
pyaudio
SpeechRecognition
pyttsx3
